backend/requirements.txt
fastapi
uvicorn
PyPDF2
google-generativeai
pydantic
